piece系　verticesfactory削除

+++ Shogi3DBattle/header/I_NaturalBufferedData.h
@@ -27,7 +27,7 @@ public:
         return S_OK;
     }
 
-    unsigned int GetSize() override {return _datas.size(); }
+    unsigned int GetSize() override {return static_cast<unsigned int>(_datas.size()); }
 
     void SetDatas(std::vector<T> datas){_datas = datas;} // データ集合セット
 };
